requirements.txt
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.1.0
cached-property==1.5.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
importlib-metadata==6.7.0
inflect==6.0.5
iniconfig==2.0.0
multidict==6.0.5
openai==1.38.0
packaging==24.0
playsound==1.2.2
pluggy==1.2.0
pydantic==1.10.17
pytest==7.4.4
python-dotenv==0.21.1
sniffio==1.3.1
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.7.1
yarl==1.9.4
zipp==3.15.0
